Just move builder methods to structureSkeleton and make it work
Refinement will come in the next commits then.
