+++ src/main/java/com/minecolonies/coremod/entity/ai/citizen/miner/EntityAIStructureMiner.java
@@ -75,6 +75,10 @@ public class EntityAIStructureMiner extends AbstractEntityAIStructure<JobMiner>
     {
         super(job);
         super.registerTargets(
+          /**
+           * If IDLE - switch to start working.
+           */
+          new AITarget(IDLE, START_WORKING),
           new AITarget(START_WORKING, this::startWorkingAtOwnBuilding),
           new AITarget(PREPARING, this::prepareForMining),
           new AITarget(MINER_SEARCHING_LADDER, this::lookForLadder),
